tray_client/requirements.txt
pystray==0.19.5
Pillow==10.4.0
requests==2.32.3
pyttsx3==2.90
pywin32==306
